--- a/include/treelite/c_api.h
+++ b/include/treelite/c_api.h
@@ -410,9 +410,31 @@ TREELITE_DLL int TreeliteFreeModel(ModelHandle handle);
  * \{
  */
 
-TREELITE_DLL int TreeliteGTILGetPredictOutputSize(ModelHandle handle, size_t num_row, size_t* out);
+/*!
+ * \brief Given a batch of data rows, query the necessary size of array to hold predictions for all
+ *        data points.
+ * \param model Treelite Model object
+ * \param num_row Number of rows in the input
+ * \param out Size of output buffer that should be allocated
+ * \return 0 for success; -1 for failure
+ */
+TREELITE_DLL int TreeliteGTILGetPredictOutputSize(ModelHandle model, size_t num_row, size_t* out);
 
-TREELITE_DLL int TreeliteGTILPredict(ModelHandle handle, const float* input, size_t num_row,
+/*!
+ * \brief Predict with a 2D dense array
+ * \param model Treelite Model object
+ * \param input The 2D data array, laid out in row-major layout
+ * \param num_row Number of rows in the data matrix.
+ * \param output Pointer to buffer to store the output. Call TreeliteGTILGetPredictOutputSize()
+ *               to get the amount of buffer you should allocate for this parameter.
+ * \param nthread number of CPU threads to use. Set <= 0 to use all CPU cores.
+ * \param pred_transform After computing the prediction score, whether to transform it.
+ * \param out_result_size Size of output. This could be smaller than
+ *                        TreeliteGTILGetPredictOutputSize() but could never be larger than
+ *                        TreeliteGTILGetPredictOutputSize().
+ * \return 0 for success; -1 for failure
+ */
+TREELITE_DLL int TreeliteGTILPredict(ModelHandle model, const float* input, size_t num_row,
                                      float* output, int nthread, int pred_transform,
                                      size_t* out_result_size);