@hcho3 2022-02-13
@trivialfis The current GTIL is also inefficient for the extremely wide (high dimensional) data. This PR does not make it worse. Creating a predictor that's efficient for the wide data is outside the scope of this PR.