@trivialfis 2022-02-13
The code looks good to me.

> The current GTIL is also inefficient for the extremely wide (high dimensional) data.

I'm just worrying that the memory usage might go up significantly. (n_features * block_size * n_threads) .  I "fixed" the issue in xgboost (found by a user) by specifying the block size to be 1 when sparsity is lower than 0.5.