@hcho3 2022-02-13
@trivialfis Thank you for the detailed review. Let me create a new GitHub issue to make a note for models trained with very wide data.